Make Mambo disk behave slightly better in the case of very large (> 4 GB) disk images. It still has issues.

+++ sys/powerpc/mambo/mambocall.h
@@ -30,6 +30,6 @@
 #ifndef _MAMBO_MAMBOCALL_H_
 #define _MAMBO_MAMBOCALL_H_
 
-int mambocall(int op, ...);
+long mambocall(int op, ...);
 
 #endif /* _MAMBO_MAMBOCALL_H_ */
